requirements.txt
aniso8601==8.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
idna==2.9
itsdangerous==1.1.0
jdatetime==3.6.2
Jinja2==2.11.1
MarkupSafe==1.1.1
PyJWT==1.7.1
pymongo==3.10.1
pytz==2019.3
redis==3.4.1
requests==2.23.0
schedule==0.6.0
six==1.14.0
urllib3==1.25.8
voluptuous==0.11.7
Werkzeug==1.0.0
mysql-connector-python==8.0.25
protobuf==3.17.3